refactor: Added a QLearner class
refactor: Added a QLearner class

A separate QLearner class is created to keep track of the Q-values.

This way, I can write a single QLearner and reuse it for many future
projects rather than having to reinvent the wheel each time.  I've
tested out this QLearner by reimplenting the smartcab to this external
class.

This is a very simple Q-learner that uses constant values for alpha
(learning rate) and gamma (discount rate.)

At a minimum, you need to pass it initial Q-values and a list of all
possible actions in order to create a QLearner object.